Move campus card to index.js
